Added player runnable and finished the commands
